box controller update
