+++ API/Controllers/BoxController.cs
@@ -2,14 +2,12 @@ using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using API.Models;
 
-
-
 namespace API.Controllers;
 
+[Route("api/[controller]")]
+[ApiController]
 public class BoxController : Controller
 {
-    //private readonly ILogger<BoxController> _logger;
-    
     private readonly Service.Service _service;
 
     public BoxController(Service.Service service)
@@ -17,5 +15,20 @@ public class BoxController : Controller
         _service = service;
     }
 
+    [HttpGet]
+    [Route("api/getAllBoxes")]
+    public IEnumerator<string> getAllBoxes()
+    {
+        //TODO Make this Method and change return type
+        return null;
+    }
+
+    [HttpGet]
+    [Route("api/box/{boxId}")]
+    public String getBoxByID([FromRoute] int boxID)
+    {
+        //TODO Make this Method and change return type
+        return "Box" + boxID;
+    }
     
 }
